增加竞赛首页数据

+++ aust-client/aust-service/src/main/java/cn/edu/aust/dto/ProblemDTO.java
@@ -25,5 +25,5 @@ public class ProblemDTO {
   private String sampleOutput;
   private String hit;
   private String nickname;
-  private Integer contestId;
+  private Long contestId;
 }
